Create index.html
Your job is to organize the code provided to you via this index.html. The flat image is to be used as a reference to recreate it as close as possible using HTML and CSS. 

You are allowed to modify the font-family, size, colours, and background images. 

You will need to use div, span, class, and id to recreate this layout.